api/requirements.txt
django==6.0
djangorestframework==3.16.1
django-cors-headers==4.9.0
python-dotenv==1.2.1
pandas==2.3.3
openpyxl==3.1.5
scikit-learn==1.7.2
redis==5.0.1
